Refactor 'price' column in pharmacy_drugs table to use float data type and set default value to 0

+++ database/migrations/2024_09_18_062516_create_pharmacy_drugs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('pharmacy_id')->constrained();
             $table->foreignId('drug_id')->constrained();
-            $table->unsignedInteger('price')->default(0);
+            $table->float('price')->default(0);
             $table->timestamps();
         });
     }
